requirements.txt
python-telegram-bot==21.4
pydantic==2.6.3
PyYAML==6.0.1
